clean up file list component
