+++ frontend/src/components/ui/file-list/events.ts
@@ -1,0 +1,3 @@
+import type { BtrixRemoveEvent } from "@/events/btrix-remove";
+
+export type FileRemoveEvent = BtrixRemoveEvent<File>;
