+++ frontend/src/components/ui/file-list/index.ts
@@ -1,0 +1,4 @@
+import "./file-list";
+import "./file-list-item";
+
+export type { FileRemoveEvent } from "./events";
