+++ frontend/src/events/btrix-remove.ts
@@ -1,0 +1,7 @@
+export type BtrixRemoveEvent<T = unknown> = CustomEvent<{ item: T }>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    "btrix-remove": BtrixRemoveEvent;
+  }
+}
