--- frontend/src/events/index.ts
@@ -1,2 +1,0 @@
-import "./btrix-change";
-import "./btrix-input";
